refactor: consolidate git utilities and resolve path ambiguity
- Move validate_git_repo, resolve_git_exclude_path, and resolve_git_dir
  from lib.rs to git.rs, where all git logic belongs. Re-export from
  lib.rs so existing call sites are unchanged.
- Extract VERSION, version_string(), and check_for_updates() from
  cli.rs to update.rs (module was already declared; now populated).
- Replace the eprintln! deprecation warning for bare local paths with
  './mydir' for local paths to avoid ambiguity with overlay references.
  Resolves the TODO comments in lib.rs and reference.rs.
